+++ internal/config/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Patch Changes
 
+- [`16ebec6`](https://github.com/thorswap/SwapKit/commit/16ebec6d5cbba777880ba24542f93a6a808c55a1) Thanks [@chillios-dev](https://github.com/chillios-dev)! - bump
+
+## null-rc.0
+
+### Patch Changes
+
+- bump
+
+## null-rc.0
+
+### Patch Changes
+
 - [#473](https://github.com/thorswap/SwapKit/pull/473) [`0864ab0`](https://github.com/thorswap/SwapKit/commit/0864ab0201cdd55ad82f27f042e38fc27d623393) Thanks [@chillios-dev](https://github.com/chillios-dev)! - Update imports
 
 - bump packages