+++ packages/wallets/xdefi/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @swapkit/wallet-xdefi
 
+## 1.0.0-rc.71
+
+### Patch Changes
+
+- [`16ebec6`](https://github.com/thorswap/SwapKit/commit/16ebec6d5cbba777880ba24542f93a6a808c55a1) Thanks [@chillios-dev](https://github.com/chillios-dev)! - bump
+
+- Updated dependencies [[`16ebec6`](https://github.com/thorswap/SwapKit/commit/16ebec6d5cbba777880ba24542f93a6a808c55a1)]:
+  - @swapkit/helpers@1.0.0-rc.56
+  - @swapkit/toolbox-cosmos@1.0.0-rc.60
+  - @swapkit/toolbox-utxo@1.0.0-rc.59
+  - @swapkit/types@1.0.0-rc.31
+  - @swapkit/toolbox-evm@1.0.0-rc.60
+
+## 1.0.0-rc.70
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.59
+  - @swapkit/helpers@1.0.0-rc.55
+  - @swapkit/toolbox-utxo@1.0.0-rc.58
+  - @swapkit/types@1.0.0-rc.30
+  - @swapkit/toolbox-evm@1.0.0-rc.59
+
 ## 1.0.0-rc.69
 
 ### Patch Changes